- webAppProject notice 모듈 refactoring 완료

+++ src/main/java/com/dev/webApp/domain/vo/NoticeVO.java
@@ -14,7 +14,7 @@ import java.sql.Timestamp;
 @AllArgsConstructor
 public class NoticeVO {
 
-    Long noticeNo;
+    Integer noticeNo;
 
     String title;
